230729 1. 플레이어 투사체 방향 완료 2. 몬스터 FSM + 아이들 애니 띄우기 완료 3. 충돌 처리 완료 (Hit, Dead State) 4. 카메라가 플레이어 따라다니는 기능 5. 오브젝트 생성 구조 변경 6. 튜토리얼 씬 더미 제거

+++ Engine_SOURCE/Lu_Animation.cpp
@@ -54,7 +54,7 @@ namespace Lu
 		SetKey(_Name);
 		m_Atlas = _Atlas;
 
-		Vector2 vResolution = Vector2(m_Atlas->GetWidth(), m_Atlas->GetHeight());
+		Vector2 vResolution = Vector2((float)m_Atlas->GetWidth(), (float)m_Atlas->GetHeight());
 
 		for (size_t i = 0; i < _ColumnLength; i++)
 		{
